requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-secret-manager==2.23.3
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
